+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## UNRELEASED
 
+* fix custom validation messages gone missing (regression from v0.9.17)
+
 ## v0.9.18
 
 * prevent infinite loops from custom validators